+++ fut/pin.py
@@ -48,12 +48,12 @@ class Pin(object):
         self.custom = {"networkAccess": "W"}  # wifi?
         # TODO?: full boot process when there is no session (boot start)
 
-        self.custom['service_plat'] = platform
-        self.s = 4  # event id  |  3 before "was sent" without session/persona/nucleus id so we can probably omit
+        self.custom['service_plat'] = platform[:3]
+        self.s = 2  # event id  |  before "was sent" without session/persona/nucleus id so we can probably omit
 
     def __ts(self):
         # TODO: add ability to random something
-        ts = datetime.now()
+        ts = datetime.utcnow()
         ts = ts.strftime('%Y-%m-%dT%H:%M:%S.%f')[:-3] + 'Z'
         return ts
 
@@ -64,9 +64,7 @@ class Pin(object):
                          "pidm": {"nucleus": self.nucleus_id},
                          "didm": {"uuid": "0"},  # what is it?
                          "ts_event": self.__ts(),
-                         "en": en},
-                'userid': self.persona_id,  # not needed before session?
-                'type': 'utas'}  # not needed before session?
+                         "en": en}}
         if self.dob:  # date of birth yyyy-mm
             data['core']['dob'] = self.dob
         if pgid:
@@ -82,14 +80,22 @@ class Pin(object):
         if end_reason:
             data['end_reason'] = end_reason
 
-        if en == 'page_view':
+        if en == 'login':
+            data['type'] = 'utas'
+            data['userid'] = self.persona_id
+        elif en == 'page_view':
             data['type'] = 'menu'
+        elif en == 'error':
+            data['server_type'] = 'utas'
+            data['errid'] = 'server_error'
+            data['type'] = 'disconnect'
+            data['sid'] = self.sid
 
         self.s += 1  # bump event id
 
         return data
 
-    def send(self, events):
+    def send(self, events, fast=False):
         time.sleep(0.5 + random() / 50)
         data = {"taxv": self.taxv,  # convert to float?
                 "tidt": self.tidt,
@@ -106,7 +112,8 @@ class Pin(object):
                 "custom": self.custom,
                 "events": events}
         # print(data)  # DEBUG
-        self.r.options(pin_url)
+        if not fast:
+            self.r.options(pin_url)
         rc = self.r.post(pin_url, data=json.dumps(data)).json()
         if rc['status'] != 'ok':
             raise FutError('PinEvent is NOT OK, probably they changed something.')